requirements.txt
opencv-python
google-generativeai
python-dotenv
moviepy